Extract tests for isFloat32Array to own file
